re-factoring of the code
