+++ lib/widgets/animations/arc_painter.dart
@@ -4,7 +4,6 @@ class ArcPainter extends CustomPainter {
   final double startAngle;
   final double sweepAngle;
   final Color color;
-  final bool isBackground;
   final double strokeWidth;
   final Gradient? gradient;
 
@@ -12,7 +11,6 @@ class ArcPainter extends CustomPainter {
     required this.startAngle,
     required this.sweepAngle,
     required this.color,
-    required this.isBackground,
     required this.strokeWidth,
     this.gradient,
   });
@@ -30,26 +28,6 @@ class ArcPainter extends CustomPainter {
       ..style = PaintingStyle.stroke
       ..strokeWidth = strokeWidth
       ..shader = gradient?.createShader(rect);
-
-    // if (isBackground) {
-    //   paint.color = color; // Solid color for the background arc
-    // } else {
-    //   // gradient for the upper arc
-    //   final gradient = LinearGradient(
-    //     colors: [
-    //       Colors.blueAccent.withOpacity(0), // Transparent start
-    //       Colors.blueAccent, // Full color
-    //       Colors.blueAccent.withOpacity(0), // Transparent end
-    //     ],
-    //     stops: const [0.0, 0.5, 1.0],
-    //   );
-
-    // Apply gradient shader
-    // paint.shader = gradient.createShader(rect);
-    // }
-
-    // const startAngle = 0.0;
-    // const sweepAngle = math.pi / 2;
 
     canvas.drawArc(rect, startAngle, sweepAngle, false, paint);
   }
