+++ lib/widgets/animations/loading_animation_list.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+import 'package:widgets_gallery/widgets/animations/arc_animation.dart';
+import 'package:widgets_gallery/widgets/animations/ft_loading_animation.dart';
+import 'package:widgets_gallery/widgets/animations/pulsating_loading_animation.dart';
+
+List<Widget> loadingAnimationList = [
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+  const FTLoadingAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const ArcAnimation(
+    size: Size(48, 48),
+    strokeWidth: 6,
+  ),
+  const PulsatingCircleAnimation(),
+];
