+++ lib/widgets/animations/transaction_animation.dart
@@ -2,30 +2,22 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:widgets_gallery/custom_page_route.dart';
+import 'package:widgets_gallery/widgets/animations/pulsating_loading_animation.dart';
 import 'package:widgets_gallery/widgets/animations/success_animation.dart';
 
-class PulsatingCircleAnimation extends StatefulWidget {
-  const PulsatingCircleAnimation({super.key});
+class TransactionAnimation extends StatefulWidget {
+  const TransactionAnimation({super.key});
 
   @override
-  State<PulsatingCircleAnimation> createState() =>
-      _PulsatingCircleAnimationState();
+  State<TransactionAnimation> createState() => _TransactionAnimationState();
 }
 
-class _PulsatingCircleAnimationState extends State<PulsatingCircleAnimation>
+class _TransactionAnimationState extends State<TransactionAnimation>
     with TickerProviderStateMixin {
-  late AnimationController _animationController;
-  late AnimationController _secondAnimationController;
   late AnimationController _textAnimationController;
   late AnimationController _columnAnimationController;
-  late Animation<double> _radiusAnimation;
-  late Animation<double> _opacityAnimation;
-  late Animation<double> _radiusAnimationForSecond;
-  late Animation<double> _opacityAnimationForSecond;
   late Animation<Offset> _textSlideAnimation;
   late Animation<double> _textOpacityAnimation;
-  bool secondCircleStarted = false;
-  bool isAnimated = false;
 
   String processingText = 'Preparing for action...';
   String processingTextSecond = 'Getting ready for a smooth transition...';
@@ -40,20 +32,6 @@ class _PulsatingCircleAnimationState extends State<PulsatingCircleAnimation>
 
   @override
   void initState() {
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(
-        milliseconds: 1500,
-      ),
-    );
-
-    _secondAnimationController = AnimationController(
-      vsync: this,
-      duration: const Duration(
-        milliseconds: 1500,
-      ),
-    );
-
     _textAnimationController = AnimationController(
       vsync: this,
       duration: const Duration(
@@ -67,60 +45,6 @@ class _PulsatingCircleAnimationState extends State<PulsatingCircleAnimation>
         milliseconds: 500,
       ),
     );
-
-    _radiusAnimation = TweenSequence([
-      TweenSequenceItem(tween: Tween<double>(begin: 10, end: 40), weight: 45),
-      TweenSequenceItem(tween: Tween<double>(begin: 40, end: 40), weight: 40),
-      TweenSequenceItem(tween: Tween<double>(begin: 40, end: 50), weight: 15),
-    ]).animate(
-      CurvedAnimation(
-        parent: _animationController,
-        curve: Curves.ease,
-      ),
-    );
-
-    _opacityAnimation = TweenSequence([
-      TweenSequenceItem(tween: Tween<double>(begin: 0, end: 1), weight: 45),
-      TweenSequenceItem(tween: Tween<double>(begin: 1, end: 1), weight: 40),
-      TweenSequenceItem(tween: Tween<double>(begin: 1, end: 0), weight: 15),
-    ]).animate(
-      CurvedAnimation(
-        parent: _animationController,
-        curve: Curves.ease,
-      ),
-    );
-
-    _animationController.repeat();
-
-    _radiusAnimationForSecond = TweenSequence([
-      TweenSequenceItem(tween: Tween<double>(begin: 10, end: 40), weight: 45),
-      TweenSequenceItem(tween: Tween<double>(begin: 40, end: 40), weight: 40),
-      TweenSequenceItem(tween: Tween<double>(begin: 40, end: 50), weight: 15),
-    ]).animate(
-      CurvedAnimation(
-        parent: _secondAnimationController,
-        curve: Curves.ease,
-      ),
-    );
-
-    _opacityAnimationForSecond = TweenSequence([
-      TweenSequenceItem(tween: Tween<double>(begin: 0, end: 1), weight: 45),
-      TweenSequenceItem(tween: Tween<double>(begin: 1, end: 1), weight: 40),
-      TweenSequenceItem(tween: Tween<double>(begin: 1, end: 0), weight: 15),
-    ]).animate(
-      CurvedAnimation(
-        parent: _secondAnimationController,
-        curve: Curves.ease,
-      ),
-    );
-
-    _animationController.addListener(() {
-      if (_radiusAnimation.value > 40 && !secondCircleStarted) {
-        secondCircleStarted = true;
-        _secondAnimationController.repeat();
-        setState(() {});
-      }
-    });
 
     _textSlideAnimation =
         Tween<Offset>(begin: Offset.zero, end: const Offset(0, -1)).animate(
@@ -183,8 +107,6 @@ class _PulsatingCircleAnimationState extends State<PulsatingCircleAnimation>
 
   @override
   void dispose() {
-    _animationController.dispose();
-    _secondAnimationController.dispose();
     _textAnimationController.dispose();
     _columnAnimationController.dispose();
     super.dispose();
@@ -204,47 +126,10 @@ class _PulsatingCircleAnimationState extends State<PulsatingCircleAnimation>
             const Spacer(
               flex: 5,
             ),
-            SizedBox(
-              width: double.infinity,
+            const SizedBox(
               height: 100,
-              child: Stack(
-                alignment: Alignment.center,
-                children: [
-                  AnimatedBuilder(
-                    animation: _animationController,
-                    builder: (context, child) {
-                      return Container(
-                        width: _radiusAnimation.value,
-                        height: _radiusAnimation.value,
-                        decoration: BoxDecoration(
-                          shape: BoxShape.circle,
-                          border: Border.all(
-                            color: Colors.white
-                                .withOpacity(_opacityAnimation.value),
-                          ),
-                        ),
-                      );
-                    },
-                  ),
-                  if (secondCircleStarted)
-                    AnimatedBuilder(
-                      animation: _secondAnimationController,
-                      builder: (context, child) {
-                        return Container(
-                          width: _radiusAnimationForSecond.value,
-                          height: _radiusAnimationForSecond.value,
-                          decoration: BoxDecoration(
-                            shape: BoxShape.circle,
-                            border: Border.all(
-                              color: Colors.white.withOpacity(
-                                _opacityAnimationForSecond.value,
-                              ),
-                            ),
-                          ),
-                        );
-                      },
-                    ),
-                ],
+              child: PulsatingCircleAnimation(
+                circleColor: Colors.white,
               ),
             ),
             const SizedBox(
